Updating Plex interactions to go through plex.direct (if possible), as Plex offers an SSL interface that can be verified.
Removing the suppression of ssl errors.

Consolidating calls to refresh section into a common function.

Adding option to clear the cached file.

Updating refresh_custom_sections to correlate the sections being refreshed with the ones that are configured in Plex, and to generate a warning if there are any missing.

Updating refresh to make a best effort to refresh all the sections even if there are some missing or fail to refresh.